Split up Name into several files, disabled propagate_generic_classifications until I can test it properly.
